Add files via upload
Here it is a C++ Program where I have used the logic of First Come First Serve and using that there are various operations on the array elements. It can push any element and display an error if the queue is overflow. 
Moreover, if the queue is underflow, it will ask the user to push in first and then and then only they are allowed to pop the elements.